refactor: extract color helpers
